Moved DynamicDependency project into a subdir so I can collapse the multiple DynDep projects under 1 parent (simplifies adding more, under the dir instead of scattered across the froot)
