+++ dev/DynamicDependency/API/DynamicDependency.vcxitems.filters
