+++ dev/DynamicDependency/API/MddCore.Architecture.h
