+++ dev/DynamicDependency/API/MddWinRT.cpp
