+++ dev/DynamicDependency/API/PackageInfo.h
