+++ dev/DynamicDependency/API/WinRTInprocModule.h
