+++ dev/DynamicDependency/API/WinRTModuleManager.cpp
