+++ dev/DynamicDependency/API/WinRTModuleManager.h
