+++ dev/DynamicDependency/API/WinRTPackage.cpp
