+++ dev/DynamicDependency/API/winrt_namespaces.h
