+++ dev/WindowsAppSDK_DLL/WindowsAppSDK_DLL.vcxproj
@@ -108,7 +108,7 @@
   <ImportGroup Label="Shared">
     <Import Project="..\AppLifecycle\AppLifecycle.vcxitems" Label="AppLifecycle" />
     <Import Project="..\PushNotifications\PushNotifications.vcxitems" Label="PushNotifications" />
-    <Import Project="..\DynamicDependency\DynamicDependency.vcxitems" Label="Shared" />
+    <Import Project="..\DynamicDependency\API\DynamicDependency.vcxitems" Label="Shared" />
     <Import Project="..\UndockedRegFreeWinRT\UndockedRegFreeWinRT.vcxitems" Label="Shared" />
     <Import Project="..\Common\Common.vcxitems" Label="Shared" />
     <Import Project="..\PowerNotifications\PowerNotifications.vcxitems" Label="Shared" />
